Moving all code for bitset functionality to a dedicated directory and library
