+++ src/core/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(utils)
+add_subdirectory(bitset)
 add_subdirectory(storage)
 add_subdirectory(compute)
