+++ src/core/SOURCE.md
@@ -2,4 +2,5 @@
 
 - `CMakeLists.txt`: CMake entry point for the core libraries.
 - `README.md`: Overview of the core subsystem structure.
+- `bitset/`: Bitset-oriented containers, filters, and their unit tests.
 - `utest_tmp_dir.h`: Shared test helper for resolving a writable temporary directory.
