+++ src/core/bitset/chunked_bits.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "roaring_ids.h"
-#include "shared_types.h"
+#include "utils/shared_types.h"
 
 #include <cstddef>
 #include <cstdint>
