+++ src/core/storage/data_merger.cpp
@@ -5,7 +5,7 @@
 #include "core/storage/data_file_layout.h"
 #include "core/storage/input_reader.h"
 #include "core/utils/log.h"
-#include "core/utils/roaring_ids.h"
+#include "core/bitset/roaring_ids.h"
 #include "core/utils/shared_consts.h"
 #include "core/utils/string_utils.h"
 #include "core/utils/timer.h"
