+++ src/core/storage/utest_data_merger.cpp
@@ -15,8 +15,8 @@
 #include "core/storage/data_merger.h"
 #include "core/storage/data_reader.h"
 #include "core/storage/input_reader.h"
-#include "core/utils/roaring_ids.h"
-#include "utest_roaring_ids_helpers.h"
+#include "core/bitset/roaring_ids.h"
+#include "core/bitset/utest_roaring_ids_helpers.h"
 #include "utest_tmp_dir.h"
 
 using namespace sketch2;
