+++ src/db/sqlite/utest_bitset_filter.cpp
@@ -8,11 +8,11 @@
 #include "sqlite3.h"
 
 #include "sketch2api/sketch2.h"
-#include "utils/bitset_filter_control.h"
-#include "utils/chunked_bits.h"
+#include "core/bitset/bitset_filter_control.h"
+#include "core/bitset/chunked_bits.h"
 #include "utils/singleton.h"
 #include "utils/shared_types.h"
-#include "utils/utest_chunked_bits_helpers.h"
+#include "core/bitset/utest_chunked_bits_helpers.h"
 
 #include <algorithm>
 #include <cstdint>
