+++ src/sketch2api/CMakeLists.txt
@@ -15,6 +15,7 @@ target_include_directories(sketch2api
 
 target_link_libraries(sketch2api
   PRIVATE
+    bitset
     compute
     stor
     utils
@@ -26,7 +27,7 @@ add_executable(utest_sketch2
   utest_sketch2api.cpp
 )
 target_include_directories(utest_sketch2 PRIVATE ${CMAKE_SOURCE_DIR}/src/core)
-target_link_libraries(utest_sketch2 PRIVATE sketch2api compute GTest::gtest Threads::Threads sketch_warnings)
+target_link_libraries(utest_sketch2 PRIVATE sketch2api compute bitset GTest::gtest Threads::Threads sketch_warnings)
 add_test(NAME utest_sketch2 COMMAND utest_sketch2)
 
 add_executable(utest_compute_sketch2api
@@ -34,5 +35,5 @@ add_executable(utest_compute_sketch2api
   utest_compute_sketch2api.cpp
 )
 target_include_directories(utest_compute_sketch2api PRIVATE ${CMAKE_SOURCE_DIR}/src/core)
-target_link_libraries(utest_compute_sketch2api PRIVATE sketch2api compute GTest::gtest Threads::Threads sketch_warnings)
+target_link_libraries(utest_compute_sketch2api PRIVATE sketch2api compute bitset GTest::gtest Threads::Threads sketch_warnings)
 add_test(NAME utest_compute_sketch2api COMMAND utest_compute_sketch2api)
