+++ src/sketch2api/internal.cpp
@@ -2,8 +2,8 @@
 
 #include "core/compute/scanner.h"
 #include "core/storage/input_generator.h"
-#include "core/utils/bitset_filter_control.h"
-#include "core/utils/chunked_bits.h"
+#include "core/bitset/bitset_filter_control.h"
+#include "core/bitset/chunked_bits.h"
 #include "core/utils/log.h"
 #include "core/utils/singleton.h"
 #include "core/utils/string_utils.h"
